+++ cs.php
@@ -7,9 +7,9 @@
  *
  *
  * @author    Marko Milost <mclion@gmail.com>
- * @copyright (c) 2006-2012 by Marko Milost / mclion
+ * @copyright (c) 2006-2012 by Marko Milost / McLion
  * @license   http://opensource.org/licenses/mit-license.php MIT
- * @version 0.2
+ * @version 0.3
  */
 
 
@@ -26,7 +26,7 @@ $cookie_show = COOKIE_NAME . 's';
 if (!isset($_COOKIE[$cookie_show])) {
     setcookie($cookie_show, uniqid('', 1), $_SERVER["REQUEST_TIME"] + 86400);
 }
-if (!check_ip($_SERVER['REMOTE_ADDR'], $_COOKIE[$cookie_show], $_GET['id'], $_SERVER["REQUEST_TIME"], 1800, 'show', $_GET['addWord'])) {
+if (!check_ip($_SERVER['REMOTE_ADDR'], $_COOKIE[$cookie_show], $_GET['id'], $_SERVER["REQUEST_TIME"], 1800, 'show', $_GET['adword'])) {
     $db->query(
 		"UPDATE 2linesads
 	   			     SET `show`=`show` + 1