+++ libs/init.php
@@ -9,15 +9,15 @@
  *
  *
  * @author    Marko Milost <mclion@gmail.com>
- * @copyright (c) 2006-2012 by Marko Milost / mclion
+ * @copyright (c) 2006-2012 by Marko Milost / McLion
  * @license   http://opensource.org/licenses/mit-license.php MIT
- * @version 0.2
+ * @version 0.3
  */
 
 header('P3P: CP="NOI NID PSD OUR IND UNI"');
 
 define('BASEDIR', '/var/www/2linesads/'); // path for includes on the server
-define('BASEURL', 'http://mclion.tuinzdaj.net/2linesads/');
+define('BASEURL', 'http://McLion.tuinzdaj.net/2linesads/');
 define('DATABASE', '2linesads');
 define('DBHOST', 'localhost');
 define('DBUSER', 'twolines');