+++ CHANGELOG.md
@@ -5,7 +5,13 @@ CHANGELOG
 
 	* This was the first version writen just to prove a point and tested for 1 year on medium traffic site selling ads and collection stats. It was ugly, but it worked well enought.
 
-* 0.2 (2006-11-07)
+* 0.2 (2012-11-07)
 
 	* Introduced AJAX + JSON in 2linesads.js. Before there was a separate PHP script that generated the javascript call. Now it does not work in IE7 and less. Not a problem, really.
 	* This is the first published version.
+
+* 0.3 (2012-11-10)
+
+	* Fixed some missmatches/typos in the code. 
+	* Keyword in 2linesads mysql table has became keywords, because it can have multiple keywords divided by |.
+